+++ src/components/view/view.component.jsx
@@ -77,7 +77,7 @@ export default class HierarchySelectorView extends React.PureComponent {
           />
         </div>
         <div className="oc-hierarchy-selector-selected-container">
-          {!this.props.standalone &&
+          {(!this.props.standalone && !this.props.hideGroupNameInput) &&
           <GroupName
             label={this.props.groupNameLabel}
             placeHolder={this.props.groupNamePlaceHolder}
@@ -249,6 +249,7 @@ HierarchySelectorView.propTypes = {
   selectedItemRenderFunction: PropTypes.func,
   standalone: PropTypes.bool,
   helpDisabled: PropTypes.bool,
+  hideGroupNameInput: PropTypes.bool,
   isClearable: PropTypes.bool,
 };
 
@@ -273,5 +274,6 @@ HierarchySelectorView.defaultProps = {
   standalone: false,
   title: '',
   helpDisabled: true,
+  hideGroupNameInput: false,
   isClearable: false,
 };